requirements.txt
streamlit
youtube-dl
openai
python-dotenv
ffmpeg-python